# openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Trivial Scanner Dashboard API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/scanner/monitor/{hostname}": {
      "get": {
        "tags": [
          "Scanner"
        ],
        "summary": "Enable Monitoring",
        "description": "Adds and enables host monitoring",
        "operationId": "enable_monitoring_scanner_monitor__hostname__get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Hostname",
              "type": "string"
            },
            "name": "hostname",
            "in": "path"
          },
          {
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            },
            "name": "authorization",
            "in": "header"
          },
          {
            "required": false,
            "schema": {
              "title": "X-Trivialscan-Account",
              "type": "string"
            },
            "name": "x-trivialscan-account",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScannerRecord"
                }
              }
            }
          },
          "401": {
            "description": "Authorization Header was sent but something was not valid (check the logs), likely signed the wrong HTTP method or forgot to sign the base64 encoded POST data"
          },
          "402": {
            "description": "Quota has been exhausted, no more monitoring is possible. Upgrade the account or stop monitoring another host"
          },
          "403": {
            "description": "Authorization Header was not sent, or dropped at a proxy (requesters issue) or the CDN (that one is our server misconfiguration)"
          },
          "500": {
            "description": "An unhandled error occurred during an AWS request for data access"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/scanner/deactivate/{hostname}": {
      "get": {
        "tags": [
          "Scanner"
        ],
        "summary": "Deactivate Monitoring",
        "description": "Adds and enables host monitoring",
        "operationId": "deactivate_monitoring_scanner_deactivate__hostname__get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Hostname",
              "type": "string"
            },
            "name": "hostname",
            "in": "path"
          },
          {
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            },
            "name": "authorization",
            "in": "header"
          },
          {
            "required": false,
            "schema": {
              "title": "X-Trivialscan-Account",
              "type": "string"
            },
            "name": "x-trivialscan-account",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScannerRecord"
                }
              }
            }
          },
          "401": {
            "description": "Authorization Header was sent but something was not valid (check the logs), likely signed the wrong HTTP method or forgot to sign the base64 encoded POST data"
          },
          "403": {
            "description": "Authorization Header was not sent, or dropped at a proxy (requesters issue) or the CDN (that one is our server misconfiguration)"
          },
          "500": {
            "description": "An unhandled error occurred during an AWS request for data access"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/scanner/queue/{hostname}": {
      "get": {
        "tags": [
          "Scanner"
        ],
        "summary": "Queue Hostname",
        "description": "Adds a host for on-demand scanning",
        "operationId": "queue_hostname_scanner_queue__hostname__get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Hostname",
              "type": "string"
            },
            "name": "hostname",
            "in": "path"
          },
          {
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            },
            "name": "authorization",
            "in": "header"
          },
          {
            "required": false,
            "schema": {
              "title": "X-Trivialscan-Account",
              "type": "string"
            },
            "name": "x-trivialscan-account",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Queue Hostname Scanner Queue  Hostname  Get",
                  "type": "boolean"
                }
              }
            }
          },
          "401": {
            "description": "Authorization Header was sent but something was not valid (check the logs), likely signed the wrong HTTP method or forgot to sign the base64 encoded POST data"
          },
          "403": {
            "description": "Authorization Header was not sent, or dropped at a proxy (requesters issue) or the CDN (that one is our server misconfiguration)"
          },
          "500": {
            "description": "An unhandled error occurred during an AWS request for data access"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Billing": {
        "title": "Billing",
        "required": [
          "product_name"
        ],
        "type": "object",
        "properties": {
          "product_name": {
            "title": "Product Name",
            "type": "string"
          },
          "is_trial": {
            "title": "Is Trial",
            "type": "boolean",
            "default": false
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "display_amount": {
            "title": "Display Amount",
            "type": "string",
            "default": "free"
          },
          "display_period": {
            "title": "Display Period",
            "type": "string"
          },
          "next_due": {
            "title": "Next Due",
            "type": "integer"
          },
          "has_invoice": {
            "title": "Has Invoice",
            "type": "boolean",
            "default": false
          }
        }
      },
...
      "Certificate": {
        "title": "Certificate",
        "required": [
          "sha1_fingerprint"
        ],
        "type": "object",
        "properties": {
          "authority_key_identifier": {
            "title": "Authority Key Identifier",
            "type": "string"
          },
          "expired": {
            "title": "Expired",
            "type": "boolean"
          },
          "expiry_status": {
            "title": "Expiry Status",
            "type": "string"
          },
          "extensions": {
            "title": "Extensions",
            "type": "array",
            "items": {},
            "default": []
          },
          "external_refs": {
            "title": "External Refs",
            "type": "object",
            "additionalProperties": {
              "maxLength": 65536,
              "minLength": 1,
              "type": "string",
              "format": "uri"
            },
            "default": {}
          },
          "is_self_signed": {
            "title": "Is Self Signed",
            "type": "boolean"
          },
          "issuer": {
            "title": "Issuer",
            "type": "string"
          },
          "known_compromised": {
            "title": "Known Compromised",
            "type": "boolean"
          },
          "md5_fingerprint": {
            "title": "Md5 Fingerprint",
            "type": "string"
          },
          "not_after": {
            "title": "Not After",
            "type": "string",
            "format": "date-time"
          },
          "not_before": {
            "title": "Not Before",
            "type": "string",
            "format": "date-time"
          },
          "public_key_curve": {
            "title": "Public Key Curve",
            "type": "string"
          },
          "public_key_exponent": {
            "title": "Public Key Exponent",
            "exclusiveMinimum": 0.0,
            "type": "integer"
          },
          "public_key_modulus": {
            "title": "Public Key Modulus",
            "exclusiveMinimum": 0.0,
            "type": "integer"
          },
          "public_key_size": {
            "title": "Public Key Size",
            "exclusiveMinimum": 0.0,
            "type": "integer"
          },
          "public_key_type": {
            "$ref": "#/components/schemas/PublicKeyType"
          },
          "revocation_crl_urls": {
            "title": "Revocation Crl Urls",
            "type": "array",
            "items": {
              "maxLength": 65536,
              "minLength": 1,
              "type": "string",
              "format": "uri"
            },
            "default": []
          },
          "san": {
            "title": "San",
            "type": "array",
            "items": {
              "type": "string"
            },
            "default": []
          },
          "serial_number": {
            "title": "Serial Number",
            "type": "string"
          },
          "serial_number_decimal": {
            "title": "Serial Number Decimal"
          },
          "serial_number_hex": {
            "title": "Serial Number Hex",
            "type": "string"
          },
          "sha1_fingerprint": {
            "title": "Sha1 Fingerprint",
            "type": "string"
          },
          "sha256_fingerprint": {
            "title": "Sha256 Fingerprint",
            "type": "string"
          },
          "signature_algorithm": {
            "title": "Signature Algorithm",
            "type": "string"
          },
          "spki_fingerprint": {
            "title": "Spki Fingerprint",
            "type": "string"
          },
          "subject": {
            "title": "Subject",
            "type": "string"
          },
          "subject_key_identifier": {
            "title": "Subject Key Identifier",
            "type": "string"
          },
          "validation_level": {
            "$ref": "#/components/schemas/ValidationLevel"
          },
          "validation_oid": {
            "title": "Validation Oid",
            "type": "string"
          },
          "version": {
            "title": "Version"
          },
          "type": {
            "$ref": "#/components/schemas/CertificateType"
          }
        }
      },
...
      "CertificateType": {
        "title": "CertificateType",
        "enum": [
          "root",
          "intermediate",
          "leaf",
          "client"
        ],
        "type": "string",
        "description": "An enumeration."
      },
...
      "ClientInfo": {
        "title": "ClientInfo",
        "type": "object",
        "properties": {
          "operating_system": {
            "title": "Operating System",
            "type": "string"
          },
          "operating_system_release": {
            "title": "Operating System Release",
            "type": "string"
          },
          "operating_system_version": {
            "title": "Operating System Version",
            "type": "string"
          },
          "architecture": {
            "title": "Architecture",
            "type": "string"
          }
        }
      },
...
      "Config": {
        "title": "Config",
        "required": [
          "defaults",
          "outputs",
          "targets"
        ],
        "type": "object",
        "properties": {
          "account_name": {
            "title": "Account Name",
            "type": "string",
            "description": "Trivial Security account name"
          },
          "client_name": {
            "title": "Client Name",
            "type": "string",
            "description": "Machine name where trivialscan CLI execcutes"
          },
          "project_name": {
            "title": "Project Name",
            "type": "string",
            "description": "Trivial Scanner project assignment for the report"
          },
          "defaults": {
            "$ref": "#/components/schemas/ConfigDefaults"
          },
          "outputs": {
            "title": "Outputs",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConfigOutput"
            }
          },
          "targets": {
            "title": "Targets",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConfigTarget"
            }
          }
        }
      },
      "ConfigDefaults": {
        "title": "ConfigDefaults",
        "required": [
          "use_sni"
        ],
        "type": "object",
        "properties": {
          "use_sni": {
            "title": "Use Sni",
            "type": "boolean"
          },
          "cafiles": {
            "title": "Cafiles",
            "type": "string"
          },
          "tmp_path_prefix": {
            "title": "Tmp Path Prefix",
            "type": "string",
            "default": "/tmp"
          },
          "http_path": {
            "title": "Http Path",
            "type": "string",
            "default": "/"
          },
          "checkpoint": {
            "title": "Checkpoint",
            "type": "boolean"
          }
        }
      },
      "ConfigOutput": {
        "title": "ConfigOutput",
        "required": [
          "type"
        ],
        "type": "object",
        "properties": {
          "type": {
            "$ref": "#/components/schemas/OutputType"
          },
          "use_icons": {
            "title": "Use Icons",
            "type": "boolean"
          },
          "when": {
            "allOf": [
              {
                "$ref": "#/components/schemas/OutputWhen"
              }
            ],
            "default": "final"
          },
          "path": {
            "title": "Path",
            "type": "string"
          }
        }
      },
      "ConfigTarget": {
        "title": "ConfigTarget",
        "required": [
          "hostname"
        ],
        "type": "object",
        "properties": {
          "hostname": {
            "title": "Hostname",
            "type": "string"
          },
          "port": {
            "title": "Port",
            "exclusiveMinimum": 0.0,
            "type": "integer",
            "default": 443
          },
          "client_certificate": {
            "title": "Client Certificate",
            "type": "string"
          },
          "http_request_paths": {
            "title": "Http Request Paths",
            "type": "array",
            "items": {
              "type": "string"
            },
            "default": [
              "/"
            ]
          }
        }
      },
...
      "Flags": {
        "title": "Flags",
        "type": "object",
        "properties": {
          "hide_progress_bars": {
            "title": "Hide Progress Bars",
            "type": "boolean"
          },
          "synchronous_only": {
            "title": "Synchronous Only",
            "type": "boolean"
          },
          "hide_banner": {
            "title": "Hide Banner",
            "type": "boolean"
          },
          "track_changes": {
            "title": "Track Changes",
            "type": "boolean"
          },
          "previous_report": {
            "title": "Previous Report",
            "type": "string"
          },
          "quiet": {
            "title": "Quiet",
            "type": "boolean"
          }
        }
      },
...
      "HTTPValidationError": {
        "title": "HTTPValidationError",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
      "Host": {
        "title": "Host",
        "required": [
          "transport"
        ],
        "type": "object",
        "properties": {
          "last_updated": {
            "title": "Last Updated",
            "type": "string",
            "format": "date-time"
          },
          "transport": {
            "$ref": "#/components/schemas/HostTransport"
          },
          "tls": {
            "$ref": "#/components/schemas/HostTLS"
          },
          "http": {
            "title": "Http",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HostHTTP"
            }
          },
          "monitoring_enabled": {
            "title": "Monitoring Enabled",
            "type": "boolean",
            "default": false
          }
        }
      },
      "HostHTTP": {
        "title": "HostHTTP",
        "required": [
          "title",
          "status_code",
          "headers",
          "body_hash"
        ],
        "type": "object",
        "properties": {
          "title": {
            "title": "Title",
            "type": "string"
          },
          "status_code": {
            "title": "Status Code",
            "maximum": 599.0,
            "minimum": 100.0,
            "type": "integer"
          },
          "headers": {
            "title": "Headers",
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "body_hash": {
            "title": "Body Hash",
            "type": "string"
          }
        }
      },
...
      "HostTLS": {
        "title": "HostTLS",
        "required": [
          "client",
          "cipher",
          "protocol",
          "session_resumption"
        ],
        "type": "object",
        "properties": {
          "certificates": {
            "title": "Certificates",
            "type": "array",
            "items": {
              "type": "string"
            },
            "default": []
          },
          "client": {
            "$ref": "#/components/schemas/HostTLSClient"
          },
          "cipher": {
            "$ref": "#/components/schemas/HostTLSCipher"
          },
          "protocol": {
            "$ref": "#/components/schemas/HostTLSProtocol"
          },
          "session_resumption": {
            "$ref": "#/components/schemas/HostTLSSessionResumption"
          }
        }
      },
...
      "HostTLSClient": {
        "title": "HostTLSClient",
        "type": "object",
        "properties": {
          "certificate_mtls_expected": {
            "title": "Certificate Mtls Expected",
            "type": "boolean",
            "default": false
          },
          "certificate_trusted": {
            "title": "Certificate Trusted",
            "type": "boolean",
            "default": false
          },
          "certificate_match": {
            "title": "Certificate Match",
            "type": "boolean",
            "default": false
          },
          "expected_client_subjects": {
            "title": "Expected Client Subjects",
            "type": "array",
            "items": {
              "type": "string"
            },
            "default": []
          }
        }
      },
      "HostTLSProtocol": {
        "title": "HostTLSProtocol",
        "required": [
          "negotiated",
          "preferred",
          "offered"
        ],
        "type": "object",
        "properties": {
          "negotiated": {
            "title": "Negotiated",
            "type": "string"
          },
          "preferred": {
            "title": "Preferred",
            "type": "string"
          },
          "offered": {
            "title": "Offered",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "HostTransport": {
        "title": "HostTransport",
        "required": [
          "hostname"
        ],
        "type": "object",
        "properties": {
          "error": {
            "title": "Error",
            "maxItems": 2,
            "minItems": 2,
            "type": "array",
            "items": [
              {
                "type": "string"
              },
              {
                "type": "string"
              }
            ]
          },
          "hostname": {
            "title": "Domain Name",
            "type": "string"
          },
          "port": {
            "title": "Port",
            "exclusiveMinimum": 0.0,
            "type": "integer",
            "default": 443
          },
          "sni_support": {
            "title": "Sni Support",
            "type": "boolean"
          },
          "peer_address": {
            "title": "Peer Address",
            "type": "string",
            "format": "ipvanyaddress"
          },
          "certificate_mtls_expected": {
            "title": "Certificate Mtls Expected",
            "type": "boolean",
            "default": false
          }
        }
      },
...
      "MemberAccountRedacted": {
        "title": "MemberAccountRedacted",
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "display": {
            "title": "Display",
            "type": "string"
          },
          "primary_email": {
            "title": "Primary Email",
            "type": "string",
            "format": "email"
          },
          "billing_email": {
            "title": "Billing Email",
            "type": "string",
            "format": "email"
          },
          "api_key": {
            "title": "Api Key",
            "type": "string"
          },
          "ip_addr": {
            "title": "Ip Addr",
            "type": "string",
            "format": "ipvanyaddress"
          },
          "user_agent": {
            "title": "User Agent",
            "type": "string"
          },
          "timestamp": {
            "title": "Timestamp",
            "type": "integer"
          },
          "billing": {
            "$ref": "#/components/schemas/Billing"
          }
        }
      },
...
      "MonitorHostname": {
        "title": "MonitorHostname",
        "required": [
          "hostname",
          "timestamp"
        ],
        "type": "object",
        "properties": {
          "hostname": {
            "title": "Hostname",
            "type": "string"
          },
          "timestamp": {
            "title": "Timestamp",
            "type": "integer"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean",
            "default": false
          }
        }
      },
...
      "PublicKeyType": {
        "title": "PublicKeyType",
        "enum": [
          "RSA",
          "DSA",
          "EC",
          "DH"
        ],
        "type": "string",
        "description": "An enumeration."
      },
...
      "ReportSummary": {
        "title": "ReportSummary",
        "required": [
          "report_id"
        ],
        "type": "object",
        "properties": {
          "generator": {
            "title": "Generator",
            "type": "string",
            "default": "trivialscan"
          },
          "version": {
            "title": "Version",
            "type": "string",
            "description": "trivialscan CLI version"
          },
          "account_name": {
            "title": "Account Name",
            "type": "string",
            "description": "Trivial Security account name"
          },
          "client_name": {
            "title": "Client Name",
            "type": "string",
            "description": "Machine name where trivialscan CLI execcutes"
          },
          "report_id": {
            "title": "Report Id",
            "type": "string"
          },
          "project_name": {
            "title": "Project Name",
            "type": "string"
          },
          "targets": {
            "title": "Targets",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Host"
            },
            "default": []
          },
          "date": {
            "title": "Date",
            "type": "string",
            "format": "date-time"
          },
          "execution_duration_seconds": {
            "title": "Execution Duration Seconds",
            "exclusiveMinimum": 0.0,
            "type": "number"
          },
          "score": {
            "title": "Score",
            "type": "integer",
            "default": 0
          },
          "results": {
            "title": "Results",
            "type": "object",
            "additionalProperties": {
              "type": "integer"
            }
          },
          "certificates": {
            "title": "Certificates",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Certificate"
            },
            "default": []
          },
          "results_uri": {
            "title": "Results Uri",
            "type": "string"
          },
          "flags": {
            "$ref": "#/components/schemas/Flags"
          },
          "config": {
            "$ref": "#/components/schemas/Config"
          },
          "client": {
            "$ref": "#/components/schemas/ClientInfo"
          },
          "type": {
            "$ref": "#/components/schemas/ScanRecordType"
          },
          "category": {
            "$ref": "#/components/schemas/ScanRecordCategory"
          },
          "is_passive": {
            "title": "Is Passive",
            "type": "boolean",
            "default": true
          }
        }
      },
...
      "ScanRecordCategory": {
        "title": "ScanRecordCategory",
        "enum": [
          "Attack Surface Monitoring",
          "Reconnaissance",
          "Public Data Sources",
          "Third Party Integration"
        ],
        "type": "string",
        "description": "An enumeration."
      },
      "ScanRecordType": {
        "title": "ScanRecordType",
        "enum": [
          "Managed Monitoring",
          "Managed On-demand",
          "Customer-managed"
        ],
        "type": "string",
        "description": "An enumeration."
      },
      "ScannerRecord": {
        "title": "ScannerRecord",
        "required": [
          "account"
        ],
        "type": "object",
        "properties": {
          "account": {
            "$ref": "#/components/schemas/MemberAccountRedacted"
          },
          "monitored_targets": {
            "title": "Monitored Targets",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MonitorHostname"
            },
            "default": []
          },
          "history": {
            "title": "History",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReportSummary"
            },
            "default": []
          }
        }
      },
...
      "ValidationError": {
        "title": "ValidationError",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "type": "object",
        "properties": {
          "loc": {
            "title": "Location",
            "type": "array",
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            }
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        }
      },
      "ValidationLevel": {
        "title": "ValidationLevel",
        "enum": [
          "Domain Validation (DV)",
          "Organization Validation (OV)",
          "Extended Validation (EV)"
        ],
        "type": "string",
        "description": "An enumeration."
      }
    }
  }
}